keep intent smaller
